Setting 로직 구현

+++ data/network/src/main/java/com/phew/network/dto/request/feed/CardFeedDto.kt
@@ -9,5 +9,5 @@ import kotlinx.serialization.Serializable
 data class CardFeedDto(
     val latitude: Double?,
     val longitude: Double?,
-    val lastId: Int?
+    val lastId: Long?
 )
